% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_domicilio.R
\name{baixa_domicilio}
\alias{baixa_domicilio}
\title{baixa dados de domicilio}
\usage{
baixa_domicilio(estado_sigla, delete.dir = TRUE)
}
\arguments{
\item{estado_sigla}{Sigla do estado}

\item{delete.dir}{apagar pasta com os arquivos baixados. Default TRUE. Modificar para FALSE caso queira manter os dados originais.}
}
\value{
dados ajustados pelos pesos do desenho amostral
}
\description{
baixa dados de domicilio
}
\examples{
# baixando os dados de Roraima
# rr<-baixa_domicilio("RR", delete.dir = FALSE)
}
