% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bota_ponto.R
\name{bota_ponto}
\alias{bota_ponto}
\title{}
\usage{
bota_ponto(x, apos = 3)
}
\arguments{
\item{x}{Caracter que vai receber ponto}

\item{apos}{Quantidade de caracter a esquerda do ponto}
}
\value{
valor com ponto
}
\description{
Função criada para auxiliar baixa_domicilio. Serve para inserir ponto que separa inteiro de decimal.
}
\examples{
bota_ponto("123456")
bota_ponto("123456", apos = 2)
bota_ponto("123456", 1)
}
